feat(edge): add get_by_guild and get_by_tag to vault-reader, bump to 0.1.21
Add guild token lookup commands to the vault-reader edge function (source
of truth at apps/kbve/edge/). Remove stale duplicate at
apps/kube/functions/manifests/custom-functions/vault-reader/ which was
never baked into the Docker image.

- get_by_guild: lookup by server_id + service name
- get_by_tag: parse "github_pat:<guild_id>" format for VaultClient compat
- Bump edge MDX to 0.1.21 to trigger new Docker image build
